Commented and fixed mistake that would equate any noise divider > 32768 with 0.

+++ Components/AY38910/AY38910.hpp
@@ -36,7 +36,7 @@ class AY38910: public ::Outputs::Filter<AY38910> {
 		int _selected_register;
 		uint8_t _registers[16], _output_registers[16];
 
-		uint16_t _tone_generator_controls[3];
+		int _tone_generator_controls[3];
 		int _channel_dividers[3];
 		int _channel_output[3];
